Add handling for invalid lines in CSV parsing
Use Callable instead of Runnable, so CsvToRdf can
get return values from thread executions, and threads
can throw exceptions.

Have thread throw an exception if a line's length is
not the same as the header line.

Clear window on invalid parsing, so as not to give the
illusion that the previously successfully conversion
can still be saved.

Signed-off-by: Cody D'Ambrosio <cjd218@lehigh.edu>